// src/shared/api/generated/models/routeGeometryDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Transit Control — proposed frontend contract
 * OpenAPI spec version: 0.1.0
 */
import type { GeometryDto } from './geometryDto';
import type { GeometryPropertiesDto } from './geometryPropertiesDto';
import type { RouteGeometryDtoType } from './routeGeometryDtoType';

export interface RouteGeometryDto {
  type: RouteGeometryDtoType;
  properties: GeometryPropertiesDto;
  geometry: GeometryDto;
}